feat: Name.com querying, addition of Premium indicator
- Adds Name.com logo, API query
- Adds a premium tag in responses

Try something like `lang.review`, and you'll see that it's premium, which means the price gets marked up significantly and doesn't work on all platforms, which means we'll need to fix that).

feat: Name.com querying, addition of Premium indicator

- Adds Name.com logo, API query
- Adds a premium tag in responses

Try something like `lang.review`, and you'll see that it's premium, which means the price gets marked up significantly and doesn't work on all platforms, which means we'll need to fix that).